tomcat/src/main/java/org/apache/catalina/session/Session.java
package org.apache.catalina.session;

import java.util.HashMap;
import java.util.Map;
import javax.annotation.Nullable;

public class Session {

    private final String id;
    private final Map<String, Object> values = new HashMap<>();

    public Session(String id) {
        this.id = id;
    }

    public String getId() {
        return id;
    }

    @Nullable
    public Object getAttribute(String name) {
        return values.get(name);
    }

    public void setAttribute(String name, Object value) {
        values.put(name, value);
    }

    public void removeAttribute(String name) {
        values.remove(name);
    }

    public void invalidate() {
        values.clear();
    }
}
